requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.3
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.3.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.5
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==38.2.0
gunicorn==21.2.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
kombu==5.6.1
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
